chore: add Compose file
Vault server generally will fail in `server` mode if a dedicated
configuration file is not available in `/vault/config` directory or if not available
as an environment variable. Through docker init container, we create this HCL
file in the `/vault/config` directory within the realm of the `vault-init` container.
This configuration file is then shared to the main `vault` container through a shared
volume `vault-init-config`. In order to achieve the provisioning `vault` container has
to wait till the `vault-init` container is successfully completed